Add method for saving authors to json.
