+++ modules/save_data.rb
@@ -1,12 +1,6 @@
-require 'json'
 
-class SaveData
-  def initialize(file_name)
-    @file_name = file_name
-  end
-
-  def save_data(data)
-    File.write(@file_name, JSON.pretty_generate(data))
-    puts 'Data saved successfully!'
+module SaveItems
+  def save_author(author)
+    @save_author_data.save_data(@author.map(&:to_hash))
   end
 end
